backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
httpx==0.28.1
pydantic==2.11.3
pydantic-settings==2.8.1
google-genai==1.11.0
redis==5.2.1
